First mighty spike at a notification plugin
This is very much a work in progress, so much so that there's not
much point even committing it for review yet, using github as my
location indepdendent storage.

This borrows heavily from the compute notifications code. The basic
idea is that there are a few classes of samples that can be had
(Fan, Temperature, Voltage, Current). Within these are potentially
multiple individual sensors and values. In all cases the volume of
a sample is always in 'Sensor Reading'.

Each class has a _transform_id and _transform_reading method which
can be used to bring the data in line with expected forms.

Questions:

* How much of the rest of the payload do we want to save, and thus
  give to the sample constructor?
* What is an event_type and how do we figure it out?
* Do we need to insure that numeric values are saved as numbers, not
  string representations thereof?
* Are the sample.nameS that are being created the correct way to
  deal with samples coming from specific devices? Or does that info
  go somewhere else? If a single instance of hardware has ten temp
  readings, how do we distinguish which reading?